backend/pip-requirements.txt
ccxt>=4.4.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
websockets>=12.0
